pyproject.toml
[project]
name = "amelia"
version = "0.1.0"
description = "A local agentic coding system supporting Dual Profiles (Work/Home)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.3",
    "pydantic-ai>=1.20.0",
    "pydantic>=2.12.4",
    "rich>=14.1.0",
    "rich-click>=1.9.4",
    "pyyaml>=6.0.2",
    "loguru>=0.7.2",
    "httpx>=0.27.2",
    "typer>=0.15.1",
]

[project.scripts]
amelia = "amelia.main:app"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["amelia"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["amelia"]
force-single-line = true
lines-after-imports = 2
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
